- moves relative import calculation into a refiner so it can be used by other languages
